+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* Fix Tooltip decoration on SideBar
+* Add more info inside README
+* Add more info inside package with comment
+
 ## 1.0.0
 
 * First Release